some more warning fixes
..could actually minimally influence the emulation, so a separate commit

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3929 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/5220intf.c
@@ -254,7 +254,7 @@ static void tms5220_update(int ch, INT16 *buffer, int length)
 
 ***********************************************************************************************/
 
-void tms5220_set_frequency(int frequency)
+void tms5220_set_frequency(double frequency)
 {
 	/* skip if output frequency is zero */
 	if (!Machine->sample_rate)
